+++ vnext/Microsoft.ReactNative.Cxx/JSI/NodeApiJsiRuntime.cpp
@@ -2061,6 +2061,8 @@ napi_value NapiJsiRuntime::GetNapiValue(const Value &value) const {
     return GetNapiValue(value.getString(*const_cast<NapiJsiRuntime *>(this)));
   } else if (value.isObject()) {
     return GetNapiValue(value.getObject(*const_cast<NapiJsiRuntime *>(this)));
+  } else if (value.isBigInt()) {
+    return GetNapiValue(value.getBigInt(*const_cast<NapiJsiRuntime *>(this)));
   } else {
     throw JSINativeException("Unexpected jsi::Value type");
   }