+++ vnext/codegen/react/components/rnwcore/EventEmitters.cpp
@@ -115,6 +115,7 @@ void AndroidSwitchEventEmitter::onChange(OnChange event) const {
   dispatchEvent("change", [event=std::move(event)](jsi::Runtime &runtime) {
     auto payload = jsi::Object(runtime);
     payload.setProperty(runtime, "value", event.value);
+payload.setProperty(runtime, "target", event.target);
     return payload;
   });
 }
@@ -122,6 +123,7 @@ void SwitchEventEmitter::onChange(OnChange event) const {
   dispatchEvent("change", [event=std::move(event)](jsi::Runtime &runtime) {
     auto payload = jsi::Object(runtime);
     payload.setProperty(runtime, "value", event.value);
+payload.setProperty(runtime, "target", event.target);
     return payload;
   });
 }